setup: work around initramfs-tools bug
A recent update (or at least recent to Raspberry Pi OS. The change appears
to have occurred between 0.142 (which works) and 0.142+deb12u1 (which does
not work) to the initramfs-tools package made it so that copying an
indirectly symlinked host file to the initramfs would create a symlink
instead of copying the file if a file with the same name was already
indirectly symlinked in the initramfs.
Work around this by using the canonical name when copying.

Also make it so debian-resizefs.sh only updates the currently used
initramfs, to speed up setup.

fixes #881